requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coverage==5.5
distlib==0.3.1
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
gunicorn==20.1.0
idna==2.10
mccabe==0.6.1
numpy==1.21.0
pandas==1.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.0
